fix(bash): native rprompt
resolves #5221
